add chunk rendering

+++ src/shaders/tile.wgsl
@@ -1,6 +1,6 @@
 struct VertexInput {
-    @location(0) position: vec2<f32>,
-    @location(1) instance_position: vec2<f32>,
+    @location(0) position: vec2<u32>,
+    @location(1) instance_position: vec2<u32>,
 }
 
 struct Screen {
@@ -14,8 +14,8 @@ struct VertexOutput {
     @builtin(position) position: vec4<f32>,
 }
 
-fn to_ndc(pixels: f32, physical_size: u32) -> f32 {
-    return pixels / f32(physical_size) * 2.0 - 1.0;
+fn to_ndc(pixels: u32, physical_size: u32) -> f32 {
+    return f32(pixels) / f32(physical_size) * 2.0 - 1.0;
 }
 
 @vertex
@@ -23,8 +23,8 @@ fn vs_main(input: VertexInput) -> VertexOutput {
     var out: VertexOutput;
 
     let pos = input.position + input.instance_position;
-    let x = to_ndc(f32(pos.x), u32(screen.size.x));
-    let y = to_ndc(f32(pos.y), u32(screen.size.y));
+    let x = to_ndc(pos.x, screen.size.x);
+    let y = to_ndc(pos.y, screen.size.y);
     out.position = vec4<f32>(x, y, 0.0, 1.0);
 
     return out;
